[리팩토링] - 코드 정리
