+++ feature/bookmark/src/main/java/com/plznoanr/lol/feature/bookmark/BookmarkScreen.kt
@@ -1,26 +1,14 @@
 package com.plznoanr.lol.feature.bookmark
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.plznoanr.lol.core.designsystem.component.EmptyBox
-import com.plznoanr.lol.core.designsystem.component.OnBottomReached
-import com.plznoanr.lol.core.designsystem.component.summoner.SummonerItem
 import com.plznoanr.lol.core.model.Summoner
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.coroutines.launch
-import timber.log.Timber
 
 @Composable
 fun BookmarkRoute(
@@ -41,31 +29,13 @@ internal fun BookmarkScreen(
     bookmarkList: PersistentList<Summoner>,
     onEvent: (Event) -> Unit,
 ) {
-    val lazyColumnState = rememberLazyListState().apply {
-        OnBottomReached {
-            Timber.d("HomeScreen call")
-            if (bookmarkList.size >= 20) {
-                onEvent(OnNextPage)
-            }
-        }
-    }
-
     if (bookmarkList.isNotEmpty()) {
-        Column {
-            TextButton(modifier = Modifier.align(Alignment.End), onClick = { onEvent(OnClear) }) {
-                Text(text = "전체 해제", color = MaterialTheme.colorScheme.secondary)
-            }
-            LazyColumn(
-                state = lazyColumnState,
-            ) {
-                items(bookmarkList) {
-                    SummonerItem(
-                        summoner = it,
-                        onBookmarked = { onEvent(OnBookmark(it.id)) }
-                    )
-                }
-            }
-        }
+        BookmarkContent(
+            bookmarkList = bookmarkList,
+            onNextPage = { onEvent(OnNextPage) },
+            onBookmark = { onEvent(OnBookmark(it)) },
+            onClear = { onEvent(OnClear) }
+        )
     } else {
         EmptyBox()
     }
